Pawan

+++ apps/backend/src/modules/user/user.controller.ts
@@ -9,33 +9,31 @@ import {
     HttpCode,
     Logger,
     Query,
+    UseGuards,
+    Req,
 } from '@nestjs/common';
-import { ApiTags, ApiOperation, ApiResponse, ApiParam, ApiQuery } from '@nestjs/swagger';
+import {
+    ApiTags,
+    ApiOperation,
+    ApiResponse,
+    ApiParam,
+    ApiQuery,
+    ApiBearerAuth,
+} from '@nestjs/swagger';
 import { UserService } from './user.service';
-import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { User } from './user.model';
 import { UUID } from 'node:crypto';
+import { JwtAuthGuard } from '../../guards/jwt-auth.guard';
+import { Request } from 'express';
 
-@ApiTags('User')
+@ApiTags('User Module')
 @Controller('user')
 export class UserController {
     private readonly logger = new Logger(UserController.name);
     constructor(private readonly userService: UserService) {}
 
-    @ApiOperation({ summary: 'Create a new user' })
-    @ApiResponse({
-        status: 201,
-        description: 'The user has been successfully created.',
-        type: User,
-    })
-    @ApiResponse({ status: 400, description: 'Bad request.' })
-    @Post()
-    create(@Body() dto: CreateUserDto) {
-        this.logger.debug('Creating a new user');
-        return this.userService.create(dto);
-    }
-
+    @Get()
     @ApiOperation({ summary: 'Get all users' })
     @ApiQuery({
         name: 'offset',
@@ -57,11 +55,11 @@ export class UserController {
         type: [User],
     })
     @ApiResponse({ status: 404, description: 'Users not found.' })
-    @Get()
     findAll(@Query('offset') offset: number, @Query('limit') limit: number) {
         return this.userService.findAll(offset, limit);
     }
 
+    @Get(':id')
     @ApiOperation({ summary: 'Get a user by ID' })
     @ApiParam({ name: 'id', description: 'User ID', type: String, format: 'uuid' })
     @ApiResponse({
@@ -70,11 +68,13 @@ export class UserController {
         type: User,
     })
     @ApiResponse({ status: 404, description: 'User not found.' })
-    @Get(':id')
     findOne(@Param('id') id: UUID) {
         return this.userService.findOne(id);
     }
 
+    @Patch(':id')
+    @UseGuards(JwtAuthGuard)
+    @ApiBearerAuth()
     @ApiOperation({ summary: 'Update a user by ID' })
     @ApiParam({ name: 'id', description: 'User ID', type: String, format: 'uuid' })
     @ApiResponse({
@@ -84,11 +84,17 @@ export class UserController {
     })
     @ApiResponse({ status: 400, description: 'Bad request.' })
     @ApiResponse({ status: 404, description: 'User not found.' })
-    @Patch(':id')
-    update(@Param('id') id: UUID, @Body() dto: UpdateUserDto) {
+    update(@Param('id') id: UUID, @Body() dto: UpdateUserDto, @Req() req: Request) {
+        if (req.user.id !== id) {
+            this.logger.error(`User ${req.user} is not allowed to update user ${id}`);
+            return;
+        }
         return this.userService.update(id, dto);
     }
 
+    @Delete(':id')
+    @UseGuards(JwtAuthGuard)
+    @ApiBearerAuth()
     @ApiOperation({ summary: 'Delete a user by ID' })
     @ApiParam({ name: 'id', description: 'User ID', type: String, format: 'uuid' })
     @ApiResponse({
@@ -97,8 +103,12 @@ export class UserController {
     })
     @ApiResponse({ status: 404, description: 'User not found.' })
     @HttpCode(204)
-    @Delete(':id')
-    remove(@Param('id') id: UUID) {
+    remove(@Param('id') id: UUID, @Req() req: Request) {
+        if (req.user.id !== id) {
+            this.logger.error(`User ${req.user} is not allowed to delete user ${id}`);
+            return;
+        }
+        this.logger.log(`User ${req.user} is deleting user ${id}`);
         return this.userService.remove(id);
     }
 }
